Camera Calibration using OpenCV
I found some code online that didn't completely work that uses opencv to do camera calibration which calculates camera parameters (including the distortion coefficient). I fixed it to the point that it works without any errors, and I started transmuting it to fit our needs.

More info in links at the top of the file, but basically after we take some checkerboard photos with our camera, we can find the camera parameters, which we can then apply to undistort any photo.